class MqttLedRing extracted
